+++ shipcat_finalizers/Cargo.toml
@@ -11,7 +11,7 @@ edition = "2018"
 [[bin]]
 doc = false
 name = "shipcat_finalizers"
-path = "src/statuscake.rs"
+path = "src/main.rs"
 
 [dependencies]
 shipcat_definitions = { path = "../shipcat_definitions" }
